Made parse interface code easier to follow by removing some redundant code

+++ agora/agora/ParseInterface.h
@@ -12,6 +12,6 @@
     + (NSArray*) browseKeyArray;
     - (void) saveNewPostToParse: (Post*) post;
     - (Post*) getFromParseIndividual: (NSString*) object_id;
-    - (NSArray*) getFromParseListByCategory: (NSString*) category AndSkipBy: (int) skip;
-    - (NSArray*) getFromParseListRecents: (int) skip;
+    - (NSArray*) getFromParseListByCategory: (NSString*) category AndSkipBy: (NSInteger) skip;
+    - (NSArray*) getFromParseListRecents: (NSInteger) skip;
 @end
